Colors: add `calculateContrastRatio` and `isAccessible`
The main change in this commit are 2 new functions `calculateContrastRatio` and `isAccessible` (+ deprecation of `isDark` and `isBright`). Function `isAccessible` is better for making decisions about color contrast when it comes to font and background because it takes the actual font and background into account (`isDark` and `isBright` are taking only one color into account completely ignoring the other).

I've shown the use-case in `ColorShowcase` inside SX Design. This commit is a first step towards a new `<Text />` for SX Design. This component should be able to render accessible color based on the background color of its parent.

adeira-source-id: 118da605386cadde7c7dca9c07107fc4bcfef4a6